+++ Section 4.3 - Example 4.2/mdp_car_rental.h
@@ -18,7 +18,7 @@ typedef struct pos {
 
 /* MDP */
 pos car_policy(int a);
-size_t next_state(pos Ss, pos Aa);
+int next_state(pos Ss, pos Aa);
 
 /* Draw in curses */
 void draw(char c);
